Move test to separate module
